% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3e.R
\name{BerekenGeleidbaarheid}
\alias{BerekenGeleidbaarheid}
\title{nu de geleidbaarheid ec25 volgens Stuyfzand en de hco3 in mequivalent/liter berekenen
Stuyfzand, P. (1987).
Een zeer nauwkeurige berekening van het elektrischgeleidingsvermogen, ter controle en aanvulling van wateranalyses:
2e versie (SWE 87.006). Retrieved from Rijswijk:}
\usage{
BerekenGeleidbaarheid(
  metveldgemiddelden = metveldgemiddelden,
  celcius = 25,
  add_bicarbonate = TRUE,
  add_phosphate = FALSE
)
}
\description{
nu de geleidbaarheid ec25 volgens Stuyfzand en de hco3 in mequivalent/liter berekenen
Stuyfzand, P. (1987).
Een zeer nauwkeurige berekening van het elektrischgeleidingsvermogen, ter controle en aanvulling van wateranalyses:
2e versie (SWE 87.006). Retrieved from Rijswijk:
}
