% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BRO_bemonsteringsprocedure}
\alias{BRO_bemonsteringsprocedure}
\title{BRO opzoektabellen}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7 rows and 4 columns.
}
\usage{
BRO_bemonsteringsprocedure
}
\description{
Opzoektabellen met informatie uit de BRO
Deze tabel bevat de geregistreerde bemonsteringsprocedures
}
\keyword{datasets}
