% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BRO_parameterlijst}
\alias{BRO_parameterlijst}
\title{BRO opzoektabellen}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 585 rows and 6 columns.
}
\usage{
BRO_parameterlijst
}
\description{
Opzoektabellen met informatie uit de BRO
Deze tabel bevat de geregistreerde parameters
}
\keyword{datasets}
