% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{filter}
\alias{filter}
\title{Filter gegevens}
\format{
een data.frame met 8 kolommen en 340 rijen.
}
\usage{
filter
}
\description{
Test dataset op basis van het LMG met daarin informatie over de
filters. Deze
dataset kan gebruikt worden voor het testen van de functies uit dit
pakket.
}
\keyword{datasets}
