% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3e.R
\name{MaakKolomMeth}
\alias{MaakKolomMeth}
\title{voorbereiding van methoden kolom voor ec25 berekening volgens Stuyfzand}
\usage{
MaakKolomMeth(
  metveldgemiddelden = dataframeuitLeesData,
  celcius = celcius,
  add_bicarbonate = add_bicarbonate,
  add_phosphate = add_phosphate
)
}
\description{
voorbereiding van methoden kolom voor ec25 berekening volgens Stuyfzand
}
