% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{metingen}
\alias{metingen}
\title{Grondwatermetingen}
\format{
een data.frame met 11 kolommen en 211792 rijen.
}
\usage{
metingen
}
\description{
Test dataset op basis van het LMG met daarin gemeten
grondwaterconcentraties Deze
dataset kan gebruikt worden voor het testen van de functies uit dit
pakket.
}
\keyword{datasets}
