% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{parameter}
\alias{parameter}
\title{Parameter gegevens}
\format{
een data.frame met 11 kolommen en 61 rijen.
}
\usage{
parameter
}
\description{
Test dataset op basis van het LMG met daarin gemeten parameters. Deze
dataset kan gebruikt worden voor het testen van de functies uit dit
pakket.
}
\keyword{datasets}
