% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{put}
\alias{put}
\title{Put gegevens}
\format{
een data.frame met 8 kolommen en 340 rijen.
}
\usage{
put
}
\description{
Test dataset op basis van het LMG met daarin gegevens over de
putten. Deze
dataset kan gebruikt worden voor het testen van de functies uit dit
pakket.
}
\keyword{datasets}
