% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0a.R
\name{QC0a}
\alias{QC0a}
\title{QC0a. Controle XY-coordinaten}
\usage{
QC0a(d_veld, d_put, d_metingen, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met putcode en XY-coordinaten voor het bemonsterde jaar X
welke vergeleken wordt met de historische XY-coordinaten doorgegeven aan de
BRO als referentie.}

\item{d_put}{BRO geregisteerde metadata behorende bij de put of putfilter.}

\item{d_metingen}{metingen bestand met monster ID's om bij afwijkingen in de
xy-coordinaten te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controleer of de geregistreerde XY-coordinaten uit het veld
overeenkomen met de in de BRO geregistreerde putcoordinaten uit
opzoekbestand.
}
\details{
De veldwerker stelt op basis van GPS-coordinaten de locatie vast.
Aan de BRO mogen de gegevens in de volgende referentiestelsels
aangeleverd worden: RD en ETR89.

Indien de coordinaten meer dan 20 meter afwijken van de oorspronkelijke
putcoordinaten, ken het concept QC oordeel verdacht toe aan het monster.

Bij putlocaties met een of meerdere andere putlocaties binnen 20 meter
afstand, vergelijk de putlocatie met de informatie meegeleverd vanuit
'de eigen organisatie' ter identificatie van de te bemonsteren put.
\itemize{
\item RIVM: vergelijk de put met de meegeleverde foto's van de put
\item Provincies: vergelijk het label van de put met meegeleverde code.
}
}
