% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0b.R
\name{QC0b}
\alias{QC0b}
\title{QC0b. Controle landgebruik}
\usage{
QC0b(d_veld, d_put, d_metingen, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met putcode en dominant landgebruik waargenomen in het veld
voor het bemonsterde jaar X welke vergeleken wordt met de informatie
doorgegeven aan de BRO als referentie.}

\item{d_put}{BRO geregisteerde metadata behorende bij de put of putfilter.}

\item{d_metingen}{metingen bestand met monster ID's om bij afwijkingen in het
landgebruik te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op verandering in landgebruik bij de putlocatie.
}
\details{
Controleer het huidige dominante landgebruikstype aan de hand van
meegeleverde veldgegevens en vergelijk dit met het bronbestand.
De landgebruiksclassificatie als gebruikt in \emph{Representativiteit KRW
Monitoringprogramma Grondwaterkwaliteit} wordt gehanteerd
(Wattel-Koekkoek et al., 2009). Zie tabel 7.3. Indien het
landgebruikstype afwijkt, ken het concept QC oordeel twijfelachtig toe
aan het monster.
}
