% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0c.R
\name{QC0c}
\alias{QC0c}
\title{QC0c. Controle beschadiging put en filter}
\usage{
QC0c(dir, d_metingen, verbose = F)
}
\arguments{
\item{dir}{directory waarin het  QC0c_beschadiging_put.csv bestand
staat}

\item{d_metingen}{metingen bestand met monster ID's om bij beschadigingen
te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op respresentatiteit van de put en het filter.
}
\details{
Controleer of de veldmedewerker beschadigingen/afwijkingen heeft
waargenomen in het veld. Zijn er beschadigingen/afwijkingen genoteerd,
ken het concept QC oordeel twijfelachtig toe aan het monster.
}
