% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0c.R
\name{QC0c_create_file}
\alias{QC0c_create_file}
\title{QC0c_create_file. Maak bestand voor Controle QC0c}
\usage{
QC0c_create_file(dir)
}
\arguments{
\item{dir}{directory waarin het bestand aangemaakt moet worden}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.

Deze functie maakt een csv bestand aan genaamd
'QC0c_beschadiging_put.csv'. Dit bestand wordt aangemaakt in de
opgegeven directory. Als het bestand al bestaat dan wordt het niet
overschreven.

Het aangemaakte bestand moet vervolgens met de hand ingevuld worden
met behulp van de informatie afkomstig van de beheerders /
veldmedewerkers.

Als het bestand correct is ingevuld dan kan het verwerkt worden
met behulp van  \code{\link[=QC0c]{QC0c()}}. Let op, \code{\link[=QC0c]{QC0c()}} verwacht de dezelfde
bestandsnaam, verander deze dus niet.}
}
\description{
Maak een csv bestand aan wat ingevuld moet worden voor de
handmatige controle van de representativiteit van de put en filter.
}
