% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0d.R
\name{QC0d}
\alias{QC0d}
\title{QC0d. Controle afdekking put}
\usage{
QC0d(dir, d_metingen, verbose = F)
}
\arguments{
\item{dir}{directory waarin het  QC0d_afdekking_put.csv bestand
staat}

\item{d_metingen}{metingen bestand met monster ID's om bij verkeerde afdekking
te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op contaminatie van het grondwater.
}
\details{
Controleer of de veldmedewerker een put zonder een goed sluitende
afdekking heeft geconstateerd. Heeft deze bijzonderheden rondom
de afdekking genoteerd, ken het concept QG oordeel twijfelachtig
toe aan het monster.
}
