% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0e.R
\name{QC0e}
\alias{QC0e}
\title{QC0e. Controle filter en filterdiepte}
\usage{
QC0e(d_veld, d_filter, d_metingen, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met putcode en filterdiepte waargenomen in het veld
voor het bemonsterde jaar X.}

\item{d_filter}{BRO geregisteerde metadata behorende bij de putcode.}

\item{d_metingen}{metingen bestand met monster ID's om bij verkeerde filterbemonstering
te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op bemonstering van het juiste filter, op de meetfout in de filterdiepte
of de aanwezigheid van sediment in het filter.
}
\details{
Als de diepte onderkant filter (okf) meer dan 50 centimeter afwijkt dan in de BRO
geregistreerde diepte, markeer de waarde als twijfelachtig.
}
