% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0f.R
\name{QC0f}
\alias{QC0f}
\title{QC0f. Controle filterdiepte}
\usage{
QC0f(d_veld, d_filter, d_metingen, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met putcode en filterdiepte waargenomen in het veld
voor het bemonsterde jaar X.}

\item{d_filter}{BRO geregisteerde metadata behorende bij de put of putcode.}

\item{d_metingen}{metingen bestand met monster ID's om bij verkeerde filterbemonstering
te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op meetfout in de filterdiepte of aanwezigheid van
sediment in het filter.
}
\details{
Indien de filterdiepte meer dan 10 centimeter afwijkt van de
geregistreerde filterdiepte, ken het concept QC oordeel
twijfelachtig toe aan het monster.
}
