% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0g.R
\name{QC0g}
\alias{QC0g}
\title{QC0g. Controle redoxklasse grondwater}
\usage{
QC0g(d_filter, d_metingen, verbose = F)
}
\arguments{
\item{d_filter}{dataframe met putcode, watertype en redoxklasses}

\item{d_metingen}{metingen uit de nieuwe meetronde}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op fouten in de redoxparameters

Controleer of de redoxklasse (bepaald volgens werkwijze in
Bijlage II QC Protocol) overeenkomt met de redoxklasse uit
het bronbestand. Wijkt de redoxklasse 1 klasse af, ken het QC oordeel
twijfelachtig toe aan het monster. Wijkt de redoxklasse meer
dan 1 klasse af, ken het QC oordeel verdacht toe.
}
