% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC1a.R
\name{QC1a}
\alias{QC1a}
\title{QC1a. Controle parameterlijst BRO}
\usage{
QC1a(d_parameter, d_metingen, verbose = F)
}
\arguments{
\item{d_parameter}{dataframe met putcode, watertype en redoxklasses}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op ID, aquocode, CASnummer, omschrijving,
eenheid en hoedanigheid.
}
\details{
Controleer of de identificatiekenmerken van de individuele
metingen overeenkomen met de parameterlijst BRO. In geval van
afwijken, ken het concept QC oordeel verdacht toe aan de betreffende
parameters.

Indien een parameter niet op de parameterlijst BRO aanwezig is, dan
dient deze te worden aangevraagd via de servicedesk BRO. De parameterlijst
wordt hiermee uitgebreid wanneer het een legitieme, nieuwe parameter betreft.
}
