% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC1b.R
\name{QC1b}
\alias{QC1b}
\title{QC1b. Controle apparaat, techniek en procedure BRO}
\usage{
QC1b(d_veld, d_parameter, d_metingen, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met veldwaarnemingen}

\item{d_parameter}{dataframe met parameter informatie}

\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
het metingen bestand met attribute van test resultaten. In de kolom
\code{oordeel} blijkt of de locatie/monster 'onverdacht' of 'verdacht' is.
}
\description{
Controle op bemonsteringsapparaat, bemonsteringsprocedure,
waardebepalingstechniek en waardebepalingsprocedure

Controleer of de bemonstering en labbepaling conform de BRO is gedaan.
}
\details{
Indien het bemonsteringsapparaat of de -procedure afwijkt van de BRO,
ken het concept QC oordeel verdacht toe aan het monster.
Indien de waardebepalingstechniek of -procedure afwijkt van de BRO,
ken het concept QC oordeel verdacht toe aan de betreffende parameters.
}
