% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC1f.R
\name{QC1f}
\alias{QC1f}
\title{QC1f. Controle negatieve waarden}
\usage{
QC1f(d_metingen, verbose = F)
}
\arguments{
\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
het metingen bestand met attribute van test resultaten. In de kolom
\code{oordeel} blijkt of de parameter 'onverdacht' of 'verdacht' is.
}
\description{
Controle op aanwezigheid van negatieve waarden
}
\details{
Indien de waarde negatief is, ken het concept
oordeel verdacht toe aan de betreffende parameter(s).
Met uitzondering van de temperatuur
}
