% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC1g.R
\name{QC1g}
\alias{QC1g}
\title{QC1g. Controle ontbreken analysegegevens}
\usage{
QC1g(d_metingen, verbose = F)
}
\arguments{
\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
het metingen bestand met attribute van test resultaten. In de kolom
\code{oordeel} blijkt of de locatie/monster 'onverdacht' of 'verdacht' is.
}
\description{
Controle op aanwezigheid van analysegegevens voor alle genomen monsters
}
\details{
Controleer of ieder monster in het bestand meetwaarden bevat van de
parameters die zijn uitgevraagd bij het laboratorium. Indien dit niet
het geval is, ken het concept oordeel ontbrekend toe aan het monster.
}
