% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC2a.R
\name{QC2a}
\alias{QC2a}
\title{QC2a. Controle toestroming filter}
\usage{
QC2a(d_veld, d_metingen, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met veldobservaties}

\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
het metingen bestand met attribute van test resultaten. In de kolom
\code{oordeel} blijkt of de locatie/monster 'onverdacht' of 'verdacht' is.
}
\description{
Controle op de toestroming van grondwater naar het filter
}
\details{
Controleer of de grondwaterstand voor het voorpompen
hoger is dan de grondwaterstand na het voorpompen.
Indien de grondwaterstand <10 cm hoger is na het voorpompen,
ken het concept oordeel twijfelachtig toe aan het monster.
Indien de grondwaterstand >10 cm hoger is na het voorpompen,
ken het concept oordeel verdacht toe aan het monster.
}
