% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC2b.R
\name{QC2b}
\alias{QC2b}
\title{QC2b. Controle conformiteit bemonsteringsprocedure}
\usage{
QC2b(dir, d_metingen, verbose = F)
}
\arguments{
\item{dir}{directory waarin het  QC2b_afwijking_bemonsteringsprocedure.csv
bestand staat}

\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
het metingen bestand met attribute van test resultaten. In de kolom
\code{oordeel} blijkt of de locatie/monster 'onverdacht' of 'verdacht' is.
}
\description{
Controle op het voldoen aan de eisen en de voorgeschreven
werkwijze ten aanzien van bemonstering.
}
\details{
Ga na of de bemonstering heeft plaatsgevonden volgens de
voorgeschreven procedure in de NTA8017. Indien afwijkingen
zijn genoteerd door de veldwerker, ken het concept QC oordeel
twijfelachtig toe aan het monster.
}
