% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC2c.R
\name{QC2c}
\alias{QC2c}
\title{QC2c. Controle veldwaarnemingen}
\usage{
QC2c(dir, d_metingen, verbose = F)
}
\arguments{
\item{dir}{directory waarin het  QC2c_veldwaarnemingen.csv bestand
staat}

\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
het metingen bestand met attribute van test resultaten. In de kolom
\code{oordeel} blijkt of de locatie/monster 'onverdacht' of 'verdacht' is.
}
\description{
Controle op veldwaarnemingen in het veld tijdens de bemonstering
}
\details{
Ga na of er bijzonderheden zijn waargenomen tijdens de bemonstering.
Indien er bijzonderheden zijn genoteerd door de veldwerker, ken het
concept oordeel twijfelachtig toe aan het monster.
}
