% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3a.R
\name{QC3a}
\alias{QC3a}
\title{QC3a. Controle datum veldonderzoek voor labanalyse}
\usage{
QC3a(d_veld, d_metingen, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met veldobservaties}

\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op de plausibiliteit van de waarden van datum
veldonderzoek en datum labanalyse
}
\details{
Ga na of de datum labanalyse na datum veldonderzoek ligt.
Indien dit niet het geval is, ken het concept oordeel
verdacht toe aan:
\itemize{
\item het gehele monster (in geval van afwijkingen bij meerdere
parameters), of
\item de betreffende parameter (in geval van afwijking bij
1 parameter)
}
}
