% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3b.R
\name{QC3b}
\alias{QC3b}
\title{QC3b. Controle registratie putfilter in de BRO}
\usage{
QC3b(d_filter, d_metingen, verbose = F)
}
\arguments{
\item{d_filter}{dataframe met putfilter informatie}

\item{d_metingen}{metingen bestand met monster ID's om bij afwijkingen in de
putcode een oordeel toe te kennen.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controleer of de betreffende putfilter is opgenomen in de BRO.
}
\details{
Ga na of het BRO-ID van het putfilter voorkomt in de lijst met
BRO-gerigistreerde putfilters. Als deze niet hierin voorkomt, ken
het concept oordeel verdacht toe aan het monster.
}
