% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3e.R
\name{QC3e}
\alias{QC3e}
\title{QC3e. Controle EC-gemeten & EC-berekend}
\usage{
QC3e(
  d_metingen,
  ph_naam = "pH",
  hco3_naam = "HCO3",
  ec_naam = "GELDHD",
  celcius = 25,
  add_bicarbonate = F,
  add_phosphate = F,
  verbose = F
)
}
\arguments{
\item{d_metingen}{dataframe met metingen}

\item{hco3_naam}{de naam van de HCO3 variable in d_metingen. Edgecase: als
deze gespecificeerd is en "HCO3" apart nog een keer voorkomt in d_metingen met
waardes die niet "NA" zijn, dan wordt het gemiddelde gebruikt van de parameters
"hco3_naam" en "HCO3"}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Vergelijking gemeten geleidbaarheid en berekende geleidbaarheid
}
\details{
Bereken de geleidbaarheid volgens de methode in Bijlage II van het protocol.

De signaleringswaarde voor monsters is delta-EC > 10 procent.
Als de delta-EC boven de signaleringswaarde ligt, ken het
concept QC oordeel twijfelachtig toe aan het monster.
}
