% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3f.R
\name{QC3f}
\alias{QC3f}
\title{QC3f. Controle pH-veld en pH-lab}
\usage{
QC3f(d_veld, d_metingen, ph_veld_naam = "pH_veld", verbose = F)
}
\arguments{
\item{d_veld}{dataframe met veldobservaties}

\item{d_metingen}{dataframe met metingen}

\item{ph_veld_naam}{character string om te gebruiken als pH
veld. Staat standaard op "pH_veld".}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Vergelijk pH-veld en pH-lab
}
\details{
De signaleringswaarde voor monsters is delta-pH >= 2 pH-eenheden.
Als de delta-pH boven de signaleringswaarde ligt, ken het
concept QC oordeel twijfelachtig toe aan het monster.
}
