% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3h.R
\name{QC3h}
\alias{QC3h}
\title{QC3h. Controle NO3 en NH4}
\usage{
QC3h(d_metingen, verbose = F)
}
\arguments{
\item{d_metingen}{metingen bestand met monster ID's om bij afwijkingen
te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controleer relatie NO3 en NH4
}
\details{
Controleer per putfilter of aan de volgende condities wordt voldaan:
\enumerate{
\item NO3 >0.5 mg/l & NH4 >2 mg/l
\item NO3 >0.5 mg/l & NH4 > NO3
}

Indien aan ten minste één van de condities wordt voldaan, ken het concept
QC oordeel twijfelachtig toe aan de betreffende parameters. Vergelijk
veldwaarden enkel met veldwaarden en labwaarden enkel met labwaarden.
}
