% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC4f.R
\name{QC4f}
\alias{QC4f}
\title{QC4f. Controle pH en HCO3}
\usage{
QC4f(d_metingen, verbose = F)
}
\arguments{
\item{d_metingen}{metingen bestand met monster ID's om bij afwijkingen
te kunnen markeren.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controleer relatie pH en HCO3
}
\details{
Controleer per putfilter of aan de volgende condities wordt voldaan:
\enumerate{
\item pH <5 & HCO3 >15 mg/l
\item pH <5.5 & HCO3 > 50 mg/l
}

Indien aan ten minste één van de condities wordt voldaan, ken het concept
QC oordeel twijfelachtig toe aan de betreffende parameters. Vergelijk
veldwaarden enkel met veldwaarden en labwaarden enkel met labwaarden.
}
