% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC5.R
\name{QC5}
\alias{QC5}
\title{QC5. Status}
\usage{
QC5(d_metingen, verbose = F)
}
\arguments{
\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F).
Staat standaard op F.}
}
\value{
metingen bestand met eindoordeel per meting.
}
\description{
In QC5 wordt aan de metingen een QC status toegekend.

Bij elke individuele meting van een parameter in het grondwatersamenstellings-
onderzoek geeft de bronhouder (of een externe partij in opdracht van de
brondhouder) een eindoordeel over de kwaliteit van de meting. Dit eindoordeel
wordt gevormd aan de hand van een voor het hele grondwatersamenstellings-
onderzoek gebruikte beoordelingsprocedure. Het eindoordeel wordt geregistreerd
in de status kwaliteitscontrole. Het is een oordeel over de kwaliteit van de
meting van de parameter, geen oordeel over het grondwatermonster als geheel.
}
\details{
Er wordt gebruik gemaakt van vier statusdefinities: goedgekeurd, afgekeurd,
onbekend en onbeslist. Zie QC5: Status van het QC protocol voor de methodiek.
}
