% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC3e.R
\name{Rossum}
\alias{Rossum}
\title{Rossum volgens Ec_voor_Patrick.docx  voor ec25 berekening}
\usage{
Rossum(z = dataframeuitMaakKolomMeth)
}
\description{
Rossum volgens Ec_voor_Patrick.docx  voor ec25 berekening
}
