% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{veld}
\alias{veld}
\title{Veldgegevens grondwater monitoring}
\format{
een data.frame met 14 kolommen en 2025 rijen.
}
\usage{
veld
}
\description{
Test dataset op basis van het LMG met daarin veldgegevens. Deze
dataset kan gebruikt worden voor het testen van de functies uit dit
pakket.
}
\keyword{datasets}
