% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC0e.R
\name{QC0e}
\alias{QC0e}
\title{QC0e. Controle filter}
\usage{
QC0e(d_veld, d_filter, d_metingen, fl = 2, verbose = F)
}
\arguments{
\item{d_veld}{dataframe met putcode en filterdiepte waargenomen in het veld
voor het bemonsterde jaar X.}

\item{d_filter}{BRO geregisteerde metadata behorende bij de putcode.}

\item{d_metingen}{metingen bestand met monster ID's om bij verkeerde filterbemonstering
te kunnen markeren.}

\item{fl}{filterlengte. Meestal is dit 1-2 meter. Standaard ingevuld is
nu 2 meter, maar dit kan aangepast worden met deze parameter.}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controle op bemonstering van het juiste filter.
}
\details{
Als de diepte onderkant filter (okf) meer dan de helft van de
filterlengte (meestal 1-2 meter) afwijkt dan in de BRO
geregistreerde diepte, markeer de waarde als verdacht.
}
