% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC4b.R
\name{QC4b}
\alias{QC4b}
\title{QC4b. Controle technisch waardebereik}
\usage{
QC4b(d_metingen, verbose = F)
}
\arguments{
\item{d_metingen}{dataframe met metingen}

\item{verbose}{of tekstuele output uit script gewenst is (T) of niet (F). Staat
standaard op F.}
}
\value{
metingen bestand met verdachte locaties/monsters.
}
\description{
Controleer of een meting fysisch niet mogelijk is.

Parameters met een fysisch waardebereik:
\itemize{
\item pH: 0 - 14 \link{-}
}
}
\details{
Indien de meetwaarde buiten het fysisch waardebereik valt, ken
de betreffende parameter het concept QC oordeel verdacht toe.
}
